refactor: use component SearchBox replace page
